pull code script

+++ command/code.sh
@@ -10,6 +10,11 @@ if [ ! -f $conf ]; then
   cp $dir/code.conf.bak $dir/code.conf
 fi
 
+if test ! -f $conf; then
+  $OUTPUT "file $conf not exists"
+  exit 1
+fi
+
 logfile=$dir/getcode.log
 rm -rf $logfile
 
@@ -21,29 +26,12 @@ fi
 codepath=$(cd $1;pwd)
 $OUTPUT "CODE PATH: $codepath \n"
 
-# 创建代码相关的目录
-logDirs=('golang/bin' 'golang/pkg' 'golang/src' 'php' 'js' 'nginx')
-for i in ${logDirs[@]}
+while read line
 do
-  codeDir="${codepath}/${i}"
-  if [ ! -d $codeDir ]; then
-    $OUTPUT "$GREEN create code dir: $codeDir $TAILS"
-    mkdir -p $codeDir
-  fi
-done
-
-if test ! -f $conf; then
-  $OUTPUT "file $conf not exists"
-  exit 1
-fi
-
-while read micodepath
-do
-  micodepath=`echo $micodepath | grep -v "#"`
-  if [ "$micodepath" = "" ];then
+  line=`echo $line | grep -v "#"`
+  if [ "$line" = "" ];then
     continue
   fi
-  #$OUTPUT "$BLUE git pull $micodepath $TAILS"
-  getcode $micodepath
+  getcode $line
 done < $conf
 $OUTPUT "\n $YELLOW pull code complete!!!! $TAILS"
